+++ lib/squib/dsl/yaml.rb
@@ -1,23 +1,23 @@
 require_relative '../args/import'
-require_relative '../import/xlsx_importer'
+require_relative '../import/Yaml_importer'
 require_relative '../errors_warnings/warn_unexpected_params'
 
 module Squib
   # DSL method. See http://squib.readthedocs.io
-  def xlsx(opts = {}, &block)
-    DSL::Xlsx.new(__callee__).run(opts, &block)
+  def yaml(opts = {}, &block)
+    DSL::Yaml.new(__callee__).run(opts, &block)
   end
-  module_function :xlsx
+  module_function :yaml
 
   class Deck
     # DSL method. See http://squib.readthedocs.io
-    def xlsx(opts = {}, &block)
-      DSL::Xlsx.new(__callee__).run(opts, &block)
+    def yaml(opts = {}, &block)
+      DSL::Yaml.new(__callee__).run(opts, &block)
     end
   end
 
   module DSL
-    class Xlsx
+    class Yaml
       include WarnUnexpectedParams
       attr_reader :dsl_method, :block
 
@@ -26,13 +26,13 @@ module Squib
       end
 
       def self.accepted_params
-        %i( file sheet strip explode )
+        %i( file data explode )
       end
 
       def run(opts,&block)
         warn_if_unexpected opts
         import_args = Args.extract_import opts
-        importer = Squib::Import::XlsxImporter.new
+        importer = Squib::Import::YamlImporter.new
         importer.import_to_dataframe(import_args, &block)
       end
     end
